Engine
created an engine, a renderer and a simulator.
the engine is responsible for the whole work of the asynchronous shit, the renderer and the simulator.
the simulator simulates logic that we need
the renderer renders logic that we need
all this (mainly engine) is fitted inside the Frame (ex Form1)

test now creates a dummy particle that is updated each frame (currently hard coded once every 64ms) to move downwards a set amount every frame