massive update
implemented textures and their loading from file
separated VAO VBO EBO and the mesh into separate classes.
added camera and the ability to move around the plane.
added support for multiples of the same object single draw call instead of a draw call for each multiple.  added a shader class.
frankly all else i dont remember because i was too busy with this and other stuff so oh well 6 days of progress in a single commit.