3D simulation implemented. Course's final project version
added support for 3D simulation. for the commit it's turned off and it defaults to 2D. and the variable changes or buttons don't work for 3D yet.
This is the version I will be using for VU ISI .NET course's final project thesis (not sure if thesis is the best word to describe what I'll be doing but hell)