input device adjustments and general adjustments
we got movement with the keyboard now
mouse now is let got or not off of the gamewindow off of focued variable.
renamed fencemesh to setupuniforms because the first one was a stupid name
introduced enumeration for the type of shader
rendering now uses maps for changing the shader program to the right one.