shadow map time (almost)
Added validation :) me being be didnt think ill need it. but i acknowledge my copium.
added lightsource component and it is responsible for lights stuff like depth buffer and such the light itself is not being rendered.
moved the UBO from the mesh to the camera because we dont use the "model" variable but instance matrices instead and all's left is the camera projection and view matrices.
added shadowmap pipeline, renderpass, imageviews.
deleted some unnecesary files
added shadowmapping hlsl files (and compiled)
changed the rasterization frag shader's light buffer to be a buffer and not uniform (forgot to change the layout name) (bug)