moving towards raytracing
adjusted some stuff to work more "generally". implemented an interface in path tracer and rasterizer for "add entity to render".
remade the buffer helper to be a static helper class so it no longer requires to be an object for access to its methods and functions.
temporarily adjusted the mesh class to work with path tracer.
cleaned up engine class.
adjusted buffer class to work with path tracing class. will later adjust so it works with both.
created path tracing class that is just for path tracing how "VulkanRenderer" (will be renamed to rasterizer) is for rasterizing.
Implemented BLAS into raytracer.
added extensions to device creation (will have to split later).
created renderer base class to work as the top object responsible for rendering and choosing a renderer. upon choosing the rendering technique that class takes over functionality.
moved bunch of functions and methods out of rasterizer class into base renderer class (will be renamed into VulkanRenderer later for clarity but that name(space) is taken right now)