Path tracing #3. last commit before drawing i promise
last commit to act as a checkpoint. fixed command buffers (image layouts are now proper) and changed the image format to rgba8unorm (storage image and swapchain)
changed the transform buffer to camera buffer (i forgot it was supposed to be the cam buffer) and now its a uniform buffer.