Raytracing Fixed & small cleanup
renamed rendererbaseclass to vulkanrenderer
renamed vulkanrederer(rasterizer) to rasterizer
and other shit renamed
added enumeration to renderers
changed miss shader to piss yellow for debug purposes.
the problem was me being stupid. in TLAS i was buffering the matrix not the BLAS reference.
Next up: doing the same for mesh class because we got different shit we need to do depending on the renderer