mega cleanup
created MCRaster and MCRaytracing as sub classes of the mesh renderer to act for different renderers.
bunch of methods from MeshComponent are now virtual. so the sub classes have their respected attributes in them.
moved features out of the contructor and into the setup method (next update ill make it virtual too)
fixed up logical device features for rasterizer.
added a differentiator for what mesh component we need in the component creation method in entity.cs
Made so the UBO now works with the rasterizer (should probably differentiate the UBOs somehow. but for now it is what it is)