UI branch
added a work in progress UI engine system. currently it is working. but there's an issue where if a UI element moves it is cut in half (?) and only the right side renders. no matter the direction.
fixed a matrix issue in mesh component to have the model matrix be reversed order.
fixed an issue where the mesh component would create a new buffer instead updating the existing one on transform update.
added a "enqueue draw commands" function that works with indexed descriptor sets.
added a new check for camera
adjusted the pipeline to work for UI related stuff. NEEDS TO BE CHANGED BACK IN ORDER TO WORK FOR GOOD OL RASTERIZER.
Added a new texture buffer creation function to create a buffer with a custom/preloaded image.
added a mesh component for UI
adjusted a failsafe check for ending command buffers on radiance cascades