light planning
added an entity manager class that will be responsible for entity lists handling.
added a vulkan control class for UI control handling (base class).
and this was the point that i realised how shit the renderer code is so imma just fucking nuke it and rewrite the entire rendering logic to make it more adaptable. good luck me