Instanced UI
yeah so since all the meshes on the screen for the UI are the same mesh with different UVs & transforms i just said fuck it instanced rendering. done in an hour with 1 crash cuz i forgor to create the sampler.

fast UI go brr. will be important when there's a fuck-ton of text on the UI. might have to do separate in world text rendering though. but that's not a concern in the near future