fixes and dev
fixed entities not running OnStart when they should start. this is a pretty bandage-esque solution as currently it is setup to update stuff on the same thread as the renderer.
added quad vertex offsets to Control class to offset vertices in shader but not change the base mesh.
fixed the issue where the px size had to be cut in half to be accurate. twas a problem with the blender mesh by default being 2m wide.