architecture
more architectural changes. figured out a system for physics, rendering & (maybe) networking to work in tandem one after another yet on different threads to not block eachother.
added a job system but as the engine stuff is not thread safe most of the time this is not used and might not be used in favor of another mechanism. some core systems will need to be a bit more robust in order to be in compliance to multithreaded workflows.
next up controls. for real this time