i dislike this
implemented the scaling control. its now scaleable on all axis accordingly disregarding if its inverted or not etc.
i dislike it cuz of the core logic. seems bandagee and not final. likely to return to this one day to optimise it (if i ever need to or out of spite).
adjusted UI collisions to trigger the object closes to the camera. dragging events stop others.
added the ability to change the cursor when needed.
adjusted UI mesh component to update control transforms instead of recreating the buffer when appropriate

next up: add it to the docker control for it to rescale all other controls accordingly when one is rescaled.
added kernel level (?) thread id printing for debug.