krc xsd generation
generating XSD for UI. this will have to work for both engine stuff and editor and game stuff. this will take in all the Vulkan control derivatives from the assembly and make an XSD for the UI editing afterwards. currently it kinda sucks as it takes in wayyy too much shit.