continuing to dev the UI system
made the registry to be config editable. Registries are defined in the "AuroraRegistries.xml" file and then added into a dictionary of registry dictionaries. the xml is driven by the XSD, just to make things a little easier.
Added a little bit of flavor to the bootstrapper to prepare the registries.
added attribute for enums to the xml editor.
added an attribute for fields of the UI controls. they can now be used to define fields including OnEnter, OnClick, Release etc.