UI development number 6-7
registry is now fully on configs.
added the rest of the actions to the xsd.
fixed adding children to containers.
made the addChild method to be overridable.
fixed docking to work with XML system.
added the ability to tag properties with "A_VulkanControlPropertyAttribute".
added dockmode member to vulkan controls.
added color setting to the controls. its not a field but a property now, so might migrate to this design pattern later on.
now all controls can have all other controls as children. use with caution kek.
added an enum type map as enums are their own types whoops.