UI dev like 69 at this point
added a mask that makes a control invisible.
added list grid control that you guessed it is a grid view.
added a function to the registry to return an asset from the type.
changed (actually more like fixed kekw stupid me) the rendering to properly be on the x/y axis and z for depth (useless). the order matters
removed vertex offset
simplified quad data to only be UVs
removed tintHover, tintClick. this will only be on CPU side. the gpu only needs the color.