what have i done #1
started working on the world structure. for serializing the entities (tbh i forgot how i got here. went from making UI, to i need to save it, to level data structure).
created an ID map for serialization. for when serialization stuff is not understandeable what is supposed to be (mostly for lists of derived classes). the name of the class get's ID'd and saved to file to understand what components, prefabs etc we got.
started working on general deserialization.
reworked/optimized a tad how the dictionary creation works. now it aligns more to what i had in mind to begin with.