XSD and XML work. new feature.
moved some attributes around
added a new attribute called "UnlistedElement" which is supposed to be an element for a control instead of it being an attribute. used for list type attributes that cant be an attribute cuz its a list lol. alongside it updated the XSD generator and XML parser to accommodate this change.
added alignment to stack panel (will probs be moved to row settings)
added summaries to attributes so its easier to understand fuck each one does
added an abstract method to containers "RecalculateLayout".